Added soft waypoints
Will attempt to make moves as the player has instructed; however, it will take a faster route if one can be found. It will NOT do this with hard waypoints.

+++ src/main/java/botmanager/frostbalance/GameNetwork.kt
@@ -9,7 +9,7 @@ import kotlin.collections.HashSet
 
 class GameNetwork(@Transient var bot: Frostbalance, var id: String) : Containable<Frostbalance>, Container {
 
-    private var turn: Int = 0
+    private var turn: Long = 0
     var worldMap: WorldMap = WorldMap(this)
     var associatedGuilds: MutableSet<GuildWrapper> = HashSet()
